+++ fabric/src/main/java/com/chefmooon/ubesdelight/data/fabric/recipe/SmeltingRecipes.java
@@ -1,8 +1,8 @@
 package com.chefmooon.ubesdelight.data.fabric.recipe;
 
 import com.chefmooon.ubesdelight.common.registry.fabric.UbesDelightItemsImpl;
-import com.chefmooon.ubesdelight.common.tag.CommonTags;
 import com.chefmooon.ubesdelight.common.utility.fabric.RecipeUtil;
+import net.fabricmc.fabric.api.tag.convention.v2.ConventionalItemTags;
 import net.minecraft.data.recipes.RecipeCategory;
 import net.minecraft.data.recipes.RecipeOutput;
 import net.minecraft.world.item.Items;
@@ -19,6 +19,6 @@ public class SmeltingRecipes {
         RecipeUtil.offerSmeltCampSmokeRecipe(UbesDelightItemsImpl.HOPIA_MUNGGO_RAW, RecipeCategory.FOOD, UbesDelightItemsImpl.HOPIA_MUNGGO, 0.5F, 200, exporter);
         RecipeUtil.offerSmeltCampSmokeRecipe(UbesDelightItemsImpl.HOPIA_UBE_RAW, RecipeCategory.FOOD, UbesDelightItemsImpl.HOPIA_UBE, 0.5F, 200, exporter);
 
-        RecipeUtil.offerSmeltCampSmokeFromTag(CommonTags.C_FOODS_MILK, UbesDelightItemsImpl.MILK_POWDER, 0.2F, 200, exporter);
+        RecipeUtil.offerSmeltCampSmokeFromTag(ConventionalItemTags.MILK_DRINKS, UbesDelightItemsImpl.MILK_POWDER, 0.2F, 200, exporter);
     }
 }